boilerplate macros
