+++ src/fem/block/element/linear/surface/triangle/mod.rs
@@ -6,6 +6,7 @@ use super::*;
 const G: usize = 1;
 const M: usize = 2;
 const N: usize = 3;
+const O: usize = 3;
 
 pub struct Triangle<C>
 {
@@ -37,7 +38,7 @@ where
     }
 }
 
-impl<'a, C> LinearElement<'a, C, G, M, N, N> for Triangle<C>
+impl<'a, C> LinearElement<'a, C, G, M, N, O> for Triangle<C>
 where
     C: Constitutive<'a>
 {
@@ -67,99 +68,4 @@ where
     }
 }
 
-impl<'a, C> LinearSurfaceElement<'a, C, G, M, N, N> for Triangle<C>
-where
-    C: Constitutive<'a>
-{
-    fn get_reference_normal(&self) -> &ReferenceNormal
-    {
-        &self.reference_normal
-    }
-}
-
-impl<'a, C> ElasticFiniteElement<'a, C, G, N> for Triangle<C>
-where
-    C: Elastic<'a>
-{
-    fn calculate_nodal_forces(&self, nodal_coordinates: &NodalCoordinates<N>) -> NodalForces<N>
-    {
-        self.calculate_nodal_forces_linear_element(nodal_coordinates)
-    }
-    fn calculate_nodal_stiffnesses(&self, nodal_coordinates: &NodalCoordinates<N>) -> NodalStiffnesses<N>
-    {
-        self.calculate_nodal_stiffnesses_linear_element(nodal_coordinates)
-    }
-}
-
-impl<'a, C> ElasticLinearElement<'a, C, G, M, N, N> for Triangle<C>
-where
-    C: Elastic<'a>
-{}
-
-impl<'a, C> HyperelasticFiniteElement<'a, C, G, N> for Triangle<C>
-where
-    C: Hyperelastic<'a>
-{
-    fn calculate_helmholtz_free_energy(&self, nodal_coordinates: &NodalCoordinates<N>) -> Scalar
-    {
-        self.calculate_helmholtz_free_energy_linear_element(nodal_coordinates)
-    }
-}
-
-impl<'a, C> HyperelasticLinearElement<'a, C, G, M, N, N> for Triangle<C>
-where
-    C: Hyperelastic<'a>
-{}
-
-impl<'a, C> ViscoelasticFiniteElement<'a, C, G, N> for Triangle<C>
-where
-    C: Viscoelastic<'a>
-{
-    fn calculate_nodal_forces(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> NodalForces<N>
-    {
-        self.calculate_nodal_forces_linear_element(nodal_coordinates, nodal_velocities)
-    }
-    fn calculate_nodal_stiffnesses(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> NodalStiffnesses<N>
-    {
-        self.calculate_nodal_stiffnesses_linear_element(nodal_coordinates, nodal_velocities)
-    }
-}
-
-impl<'a, C> ViscoelasticLinearElement<'a, C, G, M, N, N> for Triangle<C>
-where
-    C: Viscoelastic<'a>
-{}
-
-impl<'a, C> ElasticHyperviscousFiniteElement<'a, C, G, N> for Triangle<C>
-where
-    C: ElasticHyperviscous<'a>
-{
-    fn calculate_viscous_dissipation(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> Scalar
-    {
-        self.calculate_viscous_dissipation_linear_element(nodal_coordinates, nodal_velocities)
-    }
-    fn calculate_dissipation_potential(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> Scalar
-    {
-        self.calculate_dissipation_potential_linear_element(nodal_coordinates, nodal_velocities)
-    }
-}
-
-impl<'a, C> ElasticHyperviscousLinearElement<'a, C, G, M, N, N> for Triangle<C>
-where
-    C: ElasticHyperviscous<'a>
-{}
-
-impl<'a, C> HyperviscoelasticFiniteElement<'a, C, G, N> for Triangle<C>
-where
-    C: Hyperviscoelastic<'a>
-{
-    fn calculate_helmholtz_free_energy(&self, nodal_coordinates: &NodalCoordinates<N>) -> Scalar
-    {
-        self.calculate_helmholtz_free_energy_linear_element(nodal_coordinates)
-    }
-}
-
-impl<'a, C> HyperviscoelasticLinearElement<'a, C, G, M, N, N> for Triangle<C>
-where
-    C: Hyperviscoelastic<'a>
-{}
+super::linear_surface_element_boilerplate!(Triangle);
